+++ src/Utilities/outdoorwifitest.cpp
@@ -19,7 +19,7 @@ int main(int argc, char **argv) {
   ros::init(argc, argv, "outdoorwifitest");
   ros::NodeHandle nh("~");
   ros::Rate rate(50.0);
-  float CurrentTime = 0.0;
+  double CurrentTime = 0.0;
   ros::Time begin_time = ros::Time::now();
 
   px4_command::AuxiliaryState FakeAuxstate;
